requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pandas>=2.1.0
openpyxl>=3.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
faker>=22.0.0
